debt - inline `extpath.ts` into `pfs.ts`
