+++ src/vs/base/test/node/pfs/pfs.test.ts
@@ -12,7 +12,7 @@ import { randomPath } from '../../../common/extpath.js';
 import { FileAccess } from '../../../common/network.js';
 import { basename, dirname, join, sep } from '../../../common/path.js';
 import { isWindows } from '../../../common/platform.js';
-import { configureFlushOnWrite, Promises, RimRafMode, rimrafSync, SymlinkSupport, writeFileSync } from '../../../node/pfs.js';
+import { configureFlushOnWrite, Promises, realcase, realpath, realpathSync, RimRafMode, rimrafSync, SymlinkSupport, writeFileSync } from '../../../node/pfs.js';
 import { ensureNoDisposablesAreLeakedInTestSuite } from '../../common/utils.js';
 import { flakySuite, getRandomTestPath } from '../testUtils.js';
 
@@ -488,6 +488,40 @@ flakySuite('PFS', function () {
 
 		writeFileSync(testFile, largeString);
 		assert.strictEqual(fs.readFileSync(testFile).toString(), largeString);
+	});
+
+	test('realcase', async () => {
+
+		// assume case insensitive file system
+		if (process.platform === 'win32' || process.platform === 'darwin') {
+			const upper = testDir.toUpperCase();
+			const real = await realcase(upper);
+
+			if (real) { // can be null in case of permission errors
+				assert.notStrictEqual(real, upper);
+				assert.strictEqual(real.toUpperCase(), upper);
+				assert.strictEqual(real, testDir);
+			}
+		}
+
+		// linux, unix, etc. -> assume case sensitive file system
+		else {
+			let real = await realcase(testDir);
+			assert.strictEqual(real, testDir);
+
+			real = await realcase(testDir.toUpperCase());
+			assert.strictEqual(real, testDir.toUpperCase());
+		}
+	});
+
+	test('realpath', async () => {
+		const realpathVal = await realpath(testDir);
+		assert.ok(realpathVal);
+	});
+
+	test('realpathSync', () => {
+		const realpath = realpathSync(testDir);
+		assert.ok(realpath);
 	});
 
 	ensureNoDisposablesAreLeakedInTestSuite();
