+++ src/vs/workbench/api/node/extensionHostProcess.ts
@@ -12,7 +12,6 @@ import { PendingMigrationError, isCancellationError, isSigPipeError, onUnexpecte
 import { Event } from '../../../base/common/event.js';
 import * as performance from '../../../base/common/performance.js';
 import { IURITransformer } from '../../../base/common/uriIpc.js';
-import { realpath } from '../../../base/node/extpath.js';
 import { Promises } from '../../../base/node/pfs.js';
 import { IMessagePassingProtocol } from '../../../base/parts/ipc/common/ipc.js';
 import { BufferedEmitter, PersistentProtocol, ProtocolConstants } from '../../../base/parts/ipc/common/ipc.net.js';
@@ -420,7 +419,7 @@ async function startExtensionHostProcess(): Promise<void> {
 		public readonly pid = process.pid;
 		exit(code: number) { nativeExit(code); }
 		fsExists(path: string) { return Promises.exists(path); }
-		fsRealpath(path: string) { return realpath(path); }
+		fsRealpath(path: string) { return Promises.realpath(path); }
 	};
 
 	// Attempt to load uri transformer
